% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{as.data.table.fda_data}
\alias{as.data.table.fda_data}
\title{Coerce fda_data to data.table}
\usage{
as.data.table.fda_data(x, keep.rownames = FALSE, ...)
}
\arguments{
\item{x}{An \code{fda_data} object.}

\item{keep.rownames}{Ignored (for compatibility).}

\item{...}{Additional arguments (ignored).}
}
\value{
A plain \link[data.table:data.table]{data.table::data.table} without \code{fda_data} class or
metadata.
}
\description{
Coerce fda_data to data.table
}
