% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{coef.fda_model}
\alias{coef.fda_model}
\title{Extract coefficient function beta(t) from fda_model}
\usage{
\method{coef}{fda_model}(object, ...)
}
\arguments{
\item{object}{An \code{fda_model} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
A \link[data.table:data.table]{data.table::data.table} with columns \code{time}, \code{beta}, \code{se},
\code{ci_lower}, \code{ci_upper}.
}
\description{
Extract coefficient function beta(t) from fda_model
}
