% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{fitted.fda_model}
\alias{fitted.fda_model}
\title{Extract fitted curves from fda_model}
\usage{
\method{fitted}{fda_model}(object, ...)
}
\arguments{
\item{object}{An \code{fda_model} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
A \link[data.table:data.table]{data.table::data.table} with fitted curve data.
}
\description{
Extract fitted curves from fda_model
}
