% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{predict.fda_model}
\alias{predict.fda_model}
\title{Predict method for fda_model}
\usage{
\method{predict}{fda_model}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{An \code{fda_model} object.}

\item{newdata}{Optional new data for prediction.}

\item{...}{Additional arguments (passed to backend).}
}
\value{
A \link[data.table:data.table]{data.table::data.table} of predictions.
}
\description{
Predict method for fda_model
}
