% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{print.fda_basis}
\alias{print.fda_basis}
\title{Print Method for fda_basis Objects}
\usage{
\method{print}{fda_basis}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"fda_basis"}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Displays a compact summary of a B-spline basis object.
}
\examples{
basis <- bspline_basis(seq(0, 1, length.out = 50), n_knots = 8)
print(basis)

}
