% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.fda_comparison}
\alias{print.fda_comparison}
\title{Print method for fda_comparison}
\usage{
\method{print}{fda_comparison}(x, ...)
}
\arguments{
\item{x}{An \code{fda_comparison} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisible \code{x}.
}
\description{
Print method for fda_comparison
}
