% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.fda_data}
\alias{print.fda_data}
\title{Print method for fda_data}
\usage{
\method{print}{fda_data}(x, ...)
}
\arguments{
\item{x}{An \code{fda_data} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisible \code{x}.
}
\description{
Print method for fda_data
}
