% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.fda_model}
\alias{print.fda_model}
\title{Print method for fda_model}
\usage{
\method{print}{fda_model}(x, ...)
}
\arguments{
\item{x}{An \code{fda_model} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisible \code{x}.
}
\description{
Print method for fda_model
}
