% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{residuals.fda_model}
\alias{residuals.fda_model}
\title{Extract residuals from fda_model}
\usage{
\method{residuals}{fda_model}(object, ...)
}
\arguments{
\item{object}{An \code{fda_model} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Residuals (numeric vector or data.table).
}
\description{
Extract residuals from fda_model
}
