% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{summary.fda_comparison}
\alias{summary.fda_comparison}
\title{Summary method for fda_comparison}
\usage{
\method{summary}{fda_comparison}(object, ...)
}
\arguments{
\item{object}{An \code{fda_comparison} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisible metrics data.table.
}
\description{
Summary method for fda_comparison
}
