% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{vcov.fda_model}
\alias{vcov.fda_model}
\title{Extract variance-covariance information from fda_model}
\usage{
\method{vcov}{fda_model}(object, ...)
}
\arguments{
\item{object}{An \code{fda_model} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
A \link[data.table:data.table]{data.table::data.table} of variance components, or a matrix if
available in \code{extras$vcov_matrix}.
}
\description{
Extract variance-covariance information from fda_model
}
